+++ demosplan/DemosPlanCoreBundle/Logic/Survey/SurveyService.php
@@ -14,7 +14,7 @@ use demosplan\DemosPlanCoreBundle\Entity\Procedure\Procedure;
 use demosplan\DemosPlanCoreBundle\Entity\Statement\Statement;
 use demosplan\DemosPlanCoreBundle\Entity\Survey\Survey;
 use demosplan\DemosPlanCoreBundle\Logic\CoreService;
-use demosplan\DemosPlanSurveyBundle\Repository\SurveyRepository;
+use demosplan\DemosPlanCoreBundle\Repository\SurveyRepository;
 use Doctrine\ORM\OptimisticLockException;
 use Doctrine\ORM\ORMException;
